requirements.txt
requests
beautifulsoup4
numpy
Pillow
qrcode
websockets
icalendar
